% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{cleaning}
\alias{cleaning}
\title{Easily preprocess insurance data}
\usage{
cleaning(dataset)
}
\arguments{
\item{dataset}{A data frame containing insurance data.}
}
\value{
A data frame
}
\description{
Easily preprocess insurance data
}
\details{
This function allows the user to preprocess the data frame which includes
checking for null values, modifying  the column types.
}
