% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{dataset}
\alias{dataset}
\title{Medical Cost Personal Datasets}
\format{
## `who`
A data frame with 1338 rows and 7 columns:
\describe{
  \item{age}{age of primary beneficiary}
  \item{bmi}{Body mass index, providing an understanding of body, weights that are relatively high or low relative to height,objective index of body weight (kg / square meter) using the ratio of height to weight, ideally 18.5 to 24.9}
  \item{sex}{insurance contractor gender, female, male}
  \item{children}{Number of children covered by health insurance / Number of dependents}
  \item{smoker}{yes or no for the question - if the person Smokes or not }
  \item{region}{the beneficiary's residential area in the US, northeast, southeast, southwest, northwest.}
  \item{charges}{Individual medical costs billed by health insurance}
  ...
}
}
\source{
<https://www.kaggle.com/datasets/mirichoi0218/insurance/download?datasetVersionNumber=1>
}
\usage{
dataset
}
\description{
A dataset downloaded from kaggle for modelling Medical costs for
Report ...
}
\keyword{datasets}
