\name{encode}
\alias{encode}
\title{List of Encoders}
\description{
The return value of \code{inspect_map} can be used to create inputs for the following fuctions.  Refer to vignettes for examples.

\code{\link{encode_ordinal}}:  Encode Ordinal Data Into Sequential Integers

\code{\link{encode_binary}}:  Encode Binary Data Into 0 and 1

\code{\link{encode_onehot}}:  Encode categorical data by One-hot encoding
}
\keyword{manip}
