% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{To visualise the columns of data frame}
\usage{
get_plot(df, column1, column2, plot_func)
}
\arguments{
\item{df}{A data frame containing insurance data.}

\item{column1}{A column to select from dataset.}

\item{column2}{A column to select from dataset.}

\item{plot_func}{type of plot to display (its is the name of the ggplot function)}
}
\value{
displays a plot
}
\description{
To visualise the columns of data frame
}
\details{
This function allows the user to preprocess the data frame which includes
checking for null values, modifying  the column types.
}
