% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plot.R
\name{get_plot}
\alias{get_plot}
\title{To visualise the columns of data frame}
\usage{
get_plot(df, column1 = 0, column2 = 0, fill_color = 0, plot_func)
}
\arguments{
\item{df}{A data frame containing insurance data.}

\item{column1}{A column to select from dataset.}

\item{column2}{A column to select from dataset. default is}

\item{fill_color}{A column from dataset to add fill colour to the plot , If any input other than a column name is specified the default is "#bcd4e6"}

\item{plot_func}{type of plot to display (its is the name of the ggplot function) #check ggplot on what function to choose}
}
\value{
displays a plot
}
\description{
To visualise the columns of data frame
}
\details{
This function allows the user to visualise columns in the data frame which includes
bar plots  , density plots  , scatter plots and there is an option to add color/fill to show the effect of the confounding variable
}
